add a required binding for more better feedback
This bind totally makes things better by making the package bark at you
if you try to start the app without a database bind. It also means you
can start the app _with_ the bind _before_ you start the database and
the supervisor will wait until a database is up and ready.

Also adds a reconfigure hook for changing a site setting in the
database.

Signed-off-by: Robb Kidd <robb@thekidds.org>